+++ catalog/app/components/Buttons/Iconized.spec.tsx
@@ -2,6 +2,7 @@
 // import mediaQuery from 'css-mediaquery'
 import * as React from 'react'
 import renderer from 'react-test-renderer'
+import { Add as IconAdd } from '@material-ui/icons'
 
 import * as Buttons from './'
 
@@ -19,6 +20,12 @@ describe('components/Buttons/Iconized', () => {
   it('render contained', () => {
     const tree = renderer
       .create(<Buttons.Iconized icon="ac_unit" label="A" variant="contained" />)
+      .toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+  it('render with SvgIcon component', () => {
+    const tree = renderer
+      .create(<Buttons.Iconized icon={IconAdd} label="Add Item" />)
       .toJSON()
     expect(tree).toMatchSnapshot()
   })